% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldiagramR-package.R
\docType{package}
\name{modeldiagramR-package}
\alias{modeldiagramR}
\alias{modeldiagramR-package}
\title{modeldiagramR: Generate Model Diagrams for Linear Mixed Effect Models}
\description{
Generates DiagrammeR model diagrams for hierarchical linear mixed effects models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/glinse-stat/modeldiagramr}
  \item Report bugs at \url{https://github.com/glinse-stat/modeldiagramr/issues}
}

}
\author{
\strong{Maintainer}: Greta Linse \email{greta.linse@montana.edu}

Authors:
\itemize{
  \item Mark Greenwood \email{greenwood@montana.edu}
}

}
\keyword{internal}
