% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_SSPD3_data}
\alias{simulate_SSPD3_data}
\title{Simulate Split-Split-Plot Experimental Design}
\usage{
simulate_SSPD3_data()
}
\value{
A \code{data.frame} with 120 observations on 13 variables:
\describe{
\item{ID}{Observation/split-split-plot identification number, numeric}
\item{LOCATION}{One of three locations "A", "B", and "C", a blocking variable, categorical}
\item{PLOT}{Plot identification variable, one for each location and whole plot combination, numeric}
\item{REP}{Replication number, all 1 as there was no replication, integer}
\item{WHOLE_PLOT}{Treatment, whether or not the plot was irrigated "IRR_NO" for not irrigated and "IRR_YES" for irrigated, categorical}
\item{SPLIT_PLOT}{Treatment, fungicide applied to split-plot, 4 fungicides and one control, "Fung1", ..., "Fung4", and "NFung", categorical}
\item{SPLIT_SPLIT_PLOT}{Treatment, variety of bean seeded in split-split-plot, 4 varieties, "Beans1", ..., "Beans4", categorical}
\item{TRT_COMB}{Treatments combined over WHOLE_PLOT, SPLIT_PLOT, and SPLIT_SPLIT_PLOT}
\item{RESP}{Simulated response, not based on a meaningful input values, numeric}
\item{LocationF}{Factor version of LOCATION}
\item{Irrigation}{Factor version of WHOLE_PLOT treatment}
\item{Fungicide}{Factor version of SPLIT_PLOT treatment}
\item{Variety}{Factor version of SPLIT_SPLIT_PLOT treatment}
}
}
\description{
\code{simulate_SSPD3_data()} creates the data in the \code{SSPD3} example data set.
Specifically, it creates data according to the following format:
}
\examples{
SSPD3_new <- simulate_SSPD3_data()
}
\references{
Murillo D, Gezan S (2024). \emph{FielDHub: A Shiny App for Design of Experiments in Life Sciences}. R package version 1.4.2, \url{https://CRAN.R-project.org/package=FielDHub}.
}
\keyword{internal}
