% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram_function.R
\name{model_diagram}
\alias{model_diagram}
\title{Hierarchical model diagramming}
\usage{
model_diagram(
  this_model,
  this_file_path = NULL,
  this_file_type = "PNG",
  widthVal = 800,
  heightVal = 1600,
  includeSizes = TRUE,
  includeLabels = TRUE,
  orientation = "vertical",
  scale_fontsize = 1,
  shift_fixed = 0,
  shift_random = 0,
  nodeColors = md_color(diagram = "gray25", random = "gray25", fixed = "gray25"),
  nodeFillColors = md_fill(diagram = "aliceblue", random = "aliceblue", fixed =
    "darkseagreen1"),
  nodeFontColors = md_fontColor(diagram = "black", random = "black", fixed = "black")
)
}
\arguments{
\item{this_model}{Input model. Either a lme or merMod object with a nested
random effects structure.}

\item{this_file_path}{Optional. Path to a location to export the diagram.
Default is \code{NULL}.}

\item{this_file_type}{Optional. File type to export the diagram.
Default is \code{"PNG"}.}

\item{widthVal}{Optional. Width of diagram in pixels. Default is \code{800}.}

\item{heightVal}{Optional. Height of diagram in pixels. Default is \code{1600}.}

\item{includeSizes}{Optional. Include group sizes in random effect labels.
Default is \code{TRUE}.}

\item{includeLabels}{Optional. Include labels for the model diagram components.
Default is \code{TRUE}.}

\item{orientation}{Optional. Orientation of the diagram, either vertically or
horizontally. Options are \code{"vertical"}, \code{"horizontal"}. Default is \code{"vertical"}.}

\item{scale_fontsize}{Optional. Proportional font size adjustment for model
diagram component, fixed effect, and random effect labels.
Multiplies these label font sizes by the specified amount. Default is \code{1}.}

\item{shift_fixed}{Optional. Additive x axis adjustment for fixed effect labels,
only used when \code{orientation == "horizontal"}. Default is \code{0}.}

\item{shift_random}{Optional. Additive x axis adjustment for random effect labels,
only used when \code{orientation == "horizontal"}. Default is \code{0}.}

\item{nodeColors}{Optional. Function specifying the colors (\code{\link[=md_color]{md_color()}})
for the outline of the nodes. Components can be specified individually
(\code{diagram}, \code{random}, and \code{fixed}).}

\item{nodeFillColors}{Optional. Function specifying the colors (\code{\link[=md_fill]{md_fill()}})
for the fill color of the nodes. Components can be specified individually
(\code{diagram}, \code{random}, and \code{fixed}).}

\item{nodeFontColors}{Optional. Function specifying the colors (\code{\link[=md_fontColor]{md_fontColor()}})
for the font color of text in the nodes. Components can be specified
individually (\code{diagram}, \code{random}, and \code{fixed}).}
}
\value{
A rendered DiagrammeR graph as an SVG document.
}
\description{
\code{model_diagram()} takes a hierarchical nested model and returns a DiagrammeR
object visualizing the fixed and random effects structure.
}
\details{
NOTE: When including this function in an RMD document and knitting to PDF, the
graphic size variables \code{widthVal} and \code{heightVal} do not currently work. It is
recommended that instead the image is exported to a file such as a PDF and then
reimported to the document. See the examples below.
}
\examples{
# merMod object example
library(lme4)

sleepstudy_lmer <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
summary(sleepstudy_lmer)
model_diagram(sleepstudy_lmer)

# lme object example
library(nlme)

sleepstudy_lme <- lme(Reaction ~ Days, random=~Days|Subject, data=sleepstudy)
summary(sleepstudy_lme)
model_diagram(sleepstudy_lme)

# Knitting to PDF example - Don't run
\dontrun{
model_diagram(sleepstudy_lmer,
              this_file_path="sleepstudy_lmer_modeldiagram.PDF",
              this_file_type="PDF")
knitr::include_graphics("sleepstudy_lmer_modeldiagram.PDF")
              }

}
